fix: jwtMiddle -> middlewares.js
