+++ config/middlewares.js
@@ -1,6 +1,9 @@
 import multer from "multer";
 import multerS3 from "multer-s3";
 import s3 from "./s3";
+import jwt from "jsonwebtoken";
+import baseResponse from "./baseResponeStatus";
+
 
 const multerS3Uploader = multerS3({
   s3,
@@ -19,5 +22,35 @@ const middlewares = {
     storage: multerS3Uploader,
   }),
 };
+const jwtMiddleware = (req,res,next) =>{
+  // read the token from header or url
+  const token = req.headers['x-access-token'] || req.query.token;
+  // token does not exist
+  if(!token) {
+    return res.send(baseResponse.TOKEN_EMPTY)
+  }
 
+  // create a promise that decodes the token
+  const p = new Promise(
+      (resolve, reject) => {
+        jwt.verify(token, secret_config.jwtsecret , (err, verifiedToken) => {
+          if(err) reject(err);
+          resolve(verifiedToken)
+        })
+      }
+  );
+
+  // if it has failed to verify, it will return an error message
+  const onError = (error) => {
+    return res.send(baseResponse.TOKEN_VERIFICATION_FAILURE)
+  };
+  // process the promise
+  p.then((verifiedToken)=>{
+    //비밀 번호 바뀌었을 때 검증 부분 추가 할 곳
+    req.verifiedToken = verifiedToken;
+    next();
+  }).catch(onError);
+}
+
+export default jwtMiddleware;
 export default middlewares;
